HARMONY-1562: Move getCatalogLinks into the stac module and fix audit issues.
