+++ app/util/stac.ts
@@ -119,3 +119,30 @@ export async function readCatalogItems(catalogUrl: string): Promise<StacItem[]> 
 
   return items;
 }
+
+/**
+ * Return the links to the data items from a STAC catalog. Note that for most of our
+ * services the data items are under item.assets.data, but for Giovanni the links are
+ * under item.assets['Giovanni URL']. However to make it more general we return any link
+ * for an asset that includes the 'data' role in its list of roles.
+ *
+ * @param catalogItems - a list of STAC catalog items
+ * @returns a list of URLs pointing to data (may include s3/http/https URLs)
+ */
+export function getCatalogLinks(catalogItems: StacItem[]): string[] {
+  const links = [];
+  for (const item of catalogItems) {
+    if (item.assets) {
+      for (const assetName in item.assets) {
+        const asset = item.assets[assetName];
+        if (assetName === 'data' || item.assets[assetName].roles?.includes('data')) {
+          if (asset.href) {
+            links.push(asset.href);
+          }
+        }
+      }
+    }
+  }
+
+  return links;
+}
